Renamed the controller file and did some changes
