+++ server/controllers/index.js
@@ -1,5 +1,5 @@
-const users = require('./users');
-const profiles = require('./profiles');
+const users = require('./user.controller');
+const profiles = require('./profile.controller');
 
 module.exports = {
   users,
